refactor: style 파일구조 최적화
한 폴더에 style이 두 개가 있는 경우 style 디렉토리를 만들어서 관리
