+++ frontend/src/components/common/error/GlobalError.tsx
@@ -3,7 +3,7 @@ import { AxiosError, AxiosResponse } from 'axios';
 import { FallbackProps } from 'react-error-boundary';
 import { IError } from '@utils/axios';
 import { useNavigate } from 'react-router-dom';
-import * as E from './GlobalError.style';
+import * as E from './style/GlobalError.style';
 import GradientButton from '../button/color/GradientButton';
 
 const errorTemplate = Object.freeze({
