+++ frontend/src/components/common/error/style/GlobalError.style.tsx
