+++ frontend/src/components/common/error/style/NotFound.style.tsx
