+++ frontend/src/components/common/header/style/MobileHeader.style.tsx
