+++ frontend/src/components/community/common/style/WriteInfo.style.tsx
