+++ frontend/src/components/community/detail/comment/EachComment.tsx
@@ -4,7 +4,7 @@ import WriteInfo from '@components/community/common/WriteInfo';
 import Buttons from '@components/community/common/Buttons';
 import useNestedComment from '@hooks/useNestedComment';
 import confirm from '@components/common/popup/confirm/Confirm';
-import * as EA from './EachComment.style';
+import * as EA from './style/EachComment.style';
 import NestedComment from '../nestedComment/NestedComment';
 import useDeleteComment from '../../../../query/delete/useDeleteComment';
