+++ frontend/src/components/community/list/CommunityList.tsx
@@ -4,7 +4,7 @@ import OptionType from '@components/common/dropdown/OptionType';
 import useDropdown from '@hooks/useDropdown';
 import Filter from './filter/Filter';
 import CommunityListInner from './CommunityListInner';
-import Container from './CommunityList.style';
+import Container from './style/CommunityList.style';
 import Skeleton from '../loading/Skeleton';
 
 function CommunityList() {
