+++ frontend/src/components/community/list/style/CommunityList.style.tsx
