+++ frontend/src/components/community/list/style/EachArticle.style.tsx
