+++ frontend/src/components/community/write/style/EachImage.style.tsx
